rtpcap: major function refactoring
Summary:
We do 3 types of analysis based on media (network, audio, video)
and 3 types of analysis based on analyzed unit:

1. per-packet: one entry per RTP packet
2. per-frame: one entry for all RTP packets with the same timestamp (video only)
3. per-time: one entry for all RTP packets in the same period (`period_sec`)

Refactored the code to make that explicit (before it was messy to understand
which analysis was interesting).

Also moved the list->file code outside the `analyze_*` functions, which
will allow easier unit-testing later.

Test Plan:
Ran the 4 analyses ('network-time', 'audio-packet', 'video-time',
'video-frame'), and checked the results.